backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
packaging==25.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
